+++ App/Context/AppContext.js
@@ -3,7 +3,6 @@ import React, { createContext, useState } from 'react';
 export const AppContext = createContext();
 
 const AppContextProvider = ({ children }) => {
-    const [loggedInUser, setLoggedInUser] = useState(false)
     const [userLogInInfo, setUserLogInInfo] = useState(null)
     const [transactions, setTransactions] = useState([])
     const [incomeSubCatagories, setIncomeSubCatagories] = useState("")
@@ -11,7 +10,9 @@ const AppContextProvider = ({ children }) => {
     const [parentCatagoryData, setParentCatagoryData] = useState([])
     const [parentCatagory, setParentCatagory] = useState([])
 
+    const [loadingScreen, setLoadingScreen] = useState(false)
 
+    const [wrongLoginInfo, setWrongLoginInfo] = useState(false)
 
 
     const [modal, setModal] = useState(false)
@@ -19,13 +20,12 @@ const AppContextProvider = ({ children }) => {
     return (
         <>
             <AppContext.Provider value={{
-                loggedInUser, setLoggedInUser,
                 userLogInInfo, setUserLogInInfo,
                 transactions, setTransactions,
-                incomeSubCatagories, setIncomeSubCatagories,
                 parentCatagoryData, setParentCatagoryData,
                 parentCatagory, setParentCatagory,
-
+                loadingScreen, setLoadingScreen,
+                wrongLoginInfo, setWrongLoginInfo,
                 modal, setModal
             }}>
                 {children}