+++ App/Routes/DrawerNavigation.js
@@ -1,6 +1,7 @@
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import React from 'react';
-import AddSubCatagories from '../Screens/AddSubCatagories';
+import AddCatagories from '../Screens/AddCatagories';
+import UpdateSubCatagories from '../Screens/UpdateSubCatagories';
 import UserPorfile from './../Screens/userPorfile';
 import HomeStackNavigation from './HomeStackNavigation';
 
@@ -11,7 +12,8 @@ function DrawerNavigation() {
     return (
         <Drawer.Navigator initialRouteName="Home">
             <Drawer.Screen name="Home" component={HomeStackNavigation} />
-            <Drawer.Screen name="Add Sub Catagory" component={AddSubCatagories} />
+            <Drawer.Screen name="Add Sub Catagory" component={AddCatagories} />
+            <Drawer.Screen name="Update Sub Catagories" component={UpdateSubCatagories} />
             <Drawer.Screen name="Profile" component={UserPorfile} />
         </Drawer.Navigator>
     )