+++ Index.js
@@ -12,16 +12,18 @@ const Index = () => {
     const {
         parentCatagoryData, setParentCatagoryData,
         parentCatagory, setParentCatagory,
+        setLoadingScreen, setWrongLoginInfo,
+        userLogInInfo, setUserLogInInfo
     } = useContext(AppContext)
 
-   
+
 
     const initialLoginState = {
         isLoading: true,
         userToken: null,
     };
 
-    loginReducer = (prevState, action) => {
+   const loginReducer = (prevState, action) => {
         switch (action.type) {
             case 'RETRIVE_TOKEN':
                 return {
@@ -53,6 +55,7 @@ const Index = () => {
 
     const authFunctions = useMemo(() => ({
         login: async (email, password) => {
+            setLoadingScreen(true)
             let userToken;
             userToken = null;
             const loginInputs = new FormData();
@@ -61,11 +64,12 @@ const Index = () => {
 
             await axios.post("https://myaccount.accountingarif.com/api/v1/user/login", loginInputs)
                 .then(res => {
-                    
                     userToken = res.data.data.securityToken
+                    setUserLogInInfo(res.data)
                     if (userToken != null) {
                         try {
-                             AsyncStorage.setItem("userToken", userToken)
+                            AsyncStorage.setItem("userToken", userToken)
+                            AsyncStorage.setItem("userData", JSON.stringify(res.data))
                         } catch (error) {
                             console.log("🚀 ~ file: Index.js ~ line 68 ~ login: ~ error", error)
                         }
@@ -73,6 +77,7 @@ const Index = () => {
                     dispatch({ type: "LOGIN", token: userToken })
                 })
                 .catch(e => console.log(e))
+            setLoadingScreen(false)
         },
 
         regestration: () => {
@@ -83,13 +88,13 @@ const Index = () => {
         logout: async () => {
             try {
                 await AsyncStorage.removeItem("userToken");
+                await AsyncStorage.removeItem("userData");
             } catch (error) {
                 console.log("🚀 ~ file: Index.js ~ line 81 ~ logout: ~ error", error);
             }
             dispatch({ type: "LOGOUT" });
         },
     }), []);
-
 
     useEffect(() => {
         axios.get("https://myaccount.accountingarif.com/api/v1/account/parent-without-cash")
@@ -126,6 +131,8 @@ const Index = () => {
             userToken = null;
             try {
                 userToken = await AsyncStorage.getItem("userToken")
+                const userData = await AsyncStorage.getItem("userData")
+                setUserLogInInfo(JSON.parse(userData))
             } catch (error) {
                 console.log("🚀 ~ file: Index.js ~ line 124 ~ setTimeout ~ error", error);
             }